[FIX] mail: support out-of-focus 1 Message in public page
Before this commit, when receiving a new message while the tab
is out-of-focus, the title change to display "1 Message" was only
available in `/odoo`, not the discuss public page.

This commit makes it available in both contexts.
