+++ addons/mail/static/src/core/public_web/out_of_focus_service_patch.js
